lib/wechat/requirements.txt
pycryptodome
httpx